Fixed an issue which caused the app to crash if the chromium process exited before disposing.

+++ src/InfiniSwiss.CdpSharp/ChromiumExecutionContext.cs
@@ -33,7 +33,10 @@ namespace InfiniSwiss.CdpSharp
                 this.cdpCommunicator.Dispose();
             }
 
-            this.chromiumProcess?.Kill();
+            if (!chromiumProcess.HasExited)
+            {
+                this.chromiumProcess?.Kill();
+            }
 
             disposed = true;
         }
